+++ src/IntegrationTests/Tests/Commands/Attachments/CreateAttachmentTests.cs
@@ -41,7 +41,7 @@ public class CreateAttachmentTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
