+++ src/IntegrationTests/Tests/Commands/Chemicals/AddShipmentChemicalCommandTests.cs
@@ -42,7 +42,7 @@ public sealed class AddShipmentChemicalCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
@@ -133,7 +133,7 @@ public sealed class AddShipmentChemicalCommandTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
