+++ src/IntegrationTests/Tests/Commands/Chemicals/DeleteShipmentChemicalTests.cs
@@ -42,7 +42,7 @@ public sealed class DeleteShipmentChemicalTests
             WaterAmount = 3,
             WaterAmountPerHour = 0,
             Well = "test",
-            ShipmentParts = new List<int> { 1 },
+            ShipmentParts = new List<double> { 1 },
             UpdatedBy = "ABCD@equinor.com",
             UpdatedByName = "ABCD",
         };
